Moving DslBuilder models to their home files
