+++ lib/ironfan/cloud.rb
@@ -1,233 +1,209 @@
+require 'ironfan/private_key'
+
+# Step 1: Recreate most of the EC2-style calls
+# Step 2: Use this class in place of the current DslObject based one
+# Step 3: Copy/mod for VirtualBox, remove the nonsense calls
+# Step 4: ???
+# Step 5: Profit
 module Ironfan
-  module Cloud
+  module CloudDsl
+    class Base < Ironfan::DslBuilder
+      magic :bootstrap_distro, String, :default => "ubuntu10.04-gems"
+      magic :chef_client_script, String
+      magic :flavor, String
+      magic :flavor_info, Array
+      magic :image_name, String
+      magic :ssh_user, String, :default => 'root'
 
-    #
-    # Right now only one cloud provider is implemented, so the separation
-    # between `cloud` and `cloud(:ec2)` is muddy.
-    #
-    # The goal though is to allow
-    #
-    # * cloud with no predicate -- definitions that apply to all cloud
-    #   providers. If you only use one provider ever nothing stops you from
-    #   always saying `cloud`.
-    # * Declarations irrelevant to other providers are acceptable and will be ignored
-    # * Declarations that are wrong in the context of other providers (a `public_ip`
-    #   that is not available) will presumably cause a downstream error -- it's
-    #   your responsibility to overlay with provider-correct values.
-    # * There are several declarations that *could* be sensibly abstracted, but
-    #   are not. Rather than specifying `flavor 'm1.xlarge'`, I could ask for
-    #   :ram => 15, :cores => 4 or storage => 1500 and get the cheapest machine
-    #   that met or exceeded each constraint -- the default of `:price =>
-    #   :smallest` would get me a t1.micro on EC2, a 256MB on
-    #   Rackspace. Availability zones could also plausibly be parameterized.
-    #
-    # @example
-    #     # these apply regardless of cloud provider
-    #     cloud do
-    #       # this makes sense everywhere
-    #       image_name            'maverick'
-    #
-    #       # this is not offered by many providers, and its value is non-portable;
-    #       # but if you only run in one cloud there's harm in putting it here
-    #       # or overriding it.
-    #       public_ip             '1.2.3.4'
-    #
-    #       # Implemented differently across providers but its meaning is clear
-    #       security_group        :nagios
-    #
-    #       # This is harmless for the other clouds
-    #       availability_zones   ['us-east-1d']
-    #     end
-    #
-    #     # these only apply to ec2 launches.
-    #     # `ec2` is sugar for `cloud(:ec2)`.
-    #     ec2 do
-    #       spot_price_fraction   0.4
-    #     end
-    #
-    class Base < Ironfan::DslObject
-      has_keys(
-        :name, :flavor, :image_name, :image_id, :keypair,
-        :chef_client_script, :public_ip, :permanent )
-      attr_accessor :owner
+      magic :owner, Whatever
 
-      def initialize(owner, *args)
-        self.owner = owner
+      def initialize(container, *args)
+        owner     container
         super(*args)
       end
 
-      # default values to apply where no value was set
-      # @return [Hash] hash of defaults
+      def to_hash
+        Chef::Log.warn("Using to_hash is depreciated, use attributes instead")
+        attributes
+      end
+
+      # Stub to be filled by child classes
       def defaults
-        reverse_merge!({
-          :image_name         => 'natty',
-        })
       end
 
-      # The username to ssh with.
-      # @return the ssh_user if set explicitly; otherwise, the user implied by the image name, if any; or else 'root'
-      def ssh_user(val=nil)
-        from_setting_or_image_info :ssh_user, val, 'root'
+      # # TODO: Replace the lambda with an underlay from image_info?
+      # magic :image_id, String, :default => lambda { image_info[:image_id] unless image_info.nil? }
+      def image_id
+        return @image_id if @image_id
+        image_info[:image_id] unless image_info.nil?
       end
-
-      # Location of ssh private keys
-      def ssh_identity_dir(val=nil)
-        set :ssh_identity_dir, File.expand_path(val) unless val.nil?
-        @settings.include?(:ssh_identity_dir) ? @settings[:ssh_identity_dir] : Chef::Config.ec2_key_dir
-      end
-
-      # SSH identity file used for knife ssh, knife boostrap and such
-      def ssh_identity_file(val=nil)
-        set :ssh_identity_file, File.expand_path(val) unless val.nil?
-        @settings.include?(:ssh_identity_file) ? @settings[:ssh_identity_file] : File.join(ssh_identity_dir, "#{keypair}.pem")
-      end
-
-      # ID of the machine image to use.
-      # @return the image_id if set explicitly; otherwise, the id implied by the image name
-      def image_id(val=nil)
-        from_setting_or_image_info :image_id, val
-      end
-
-      # Distribution knife should target when bootstrapping an instance
-      # @return the bootstrap_distro if set explicitly; otherwise, the bootstrap_distro implied by the image name
-      def bootstrap_distro(val=nil)
-        from_setting_or_image_info :bootstrap_distro, val, "ubuntu10.04-gems"
-      end
-
-      def validation_key
-        IO.read(Chef::Config.validation_key) rescue ''
-      end
-
-      # The instance price, drawn from the compute flavor's info
-      def price
-        flavor_info[:price]
-      end
-
-      # The instance bitness, drawn from the compute flavor's info
-      def bits
-        flavor_info[:bits]
-      end
-
-    protected
-      # If value was explicitly set, use that; if the Chef::Config[:ec2_image_info] implies a value use that; otherwise use the default
-      def from_setting_or_image_info(key, val=nil, default=nil)
-        @settings[key] = val unless val.nil?
-        return @settings[key]  if @settings.include?(key)
-        return image_info[key] unless image_info.nil?
-        return default       # otherwise
-      end
+## TODO: Replace with code that will assume ssh_identity_dir if ssh_identity_file isn't absolutely pathed
+#       # SSH identity file used for knife ssh, knife boostrap and such
+#       def ssh_identity_file(val=nil)
+#         set :ssh_identity_file, File.expand_path(val) unless val.nil?
+#         @settings.include?(:ssh_identity_file) ? @settings[:ssh_identity_file] : File.join(ssh_identity_dir, "#{keypair}.pem")
+#       end
     end
-
+    
     class Ec2 < Base
-      has_keys(
-        :region, :availability_zones, :backing,
-        :spot_price, :spot_price_fraction,
-        :user_data, :security_groups,
-        :monitoring, :placement_group,
-        :vpc, :subnet
-        )
-
-      def initialize(*args)
-        super *args
-        name :ec2 # cloud provider name
-        @settings[:security_groups]      ||= Mash.new
-        @settings[:user_data]            ||= Mash.new
-      end
-
       #
-      # Sets some defaults for amazon cloud usage, and registers the root volume
+      # To add to this list, use this snippet:
       #
-      def defaults
-        owner.volume(:root).reverse_merge!({
-            :device      => '/dev/sda1',
-            :mount_point => '/',
-            :mountable   => false,
-          })
-        self.reverse_merge!({
-            :availability_zones => ['us-east-1d'],
-            :backing            => 'ebs',
-            :flavor             => 't1.micro',
-          })
-        super
-      end
+      #     Chef::Config[:ec2_image_info] ||= {}
+      #     Chef::Config[:ec2_image_info].merge!({
+      #       # ... lines like the below
+      #     })
+      #
+      # in your knife.rb or whereever. We'll notice that it exists and add to it, rather than clobbering it.
+      #
+      Chef::Config[:ec2_image_info] ||= {}
+      Chef::Config[:ec2_image_info].merge!({
 
-      # adds a security group to the cloud instance
-      def security_group(sg_name, hsh={}, &block)
-        sg_name = sg_name.to_s
-        security_groups[sg_name] ||= Ironfan::Cloud::SecurityGroup.new(self, sg_name)
-        security_groups[sg_name].configure(hsh, &block)
-        security_groups[sg_name]
-      end
+        #
+        # Lucid (Ubuntu 9.10)
+        #
+        %w[us-east-1             64-bit  instance        karmic     ] => { :image_id => 'ami-55739e3c', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[us-east-1             32-bit  instance        karmic     ] => { :image_id => 'ami-bb709dd2', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[us-west-1             64-bit  instance        karmic     ] => { :image_id => 'ami-cb2e7f8e', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[us-west-1             32-bit  instance        karmic     ] => { :image_id => 'ami-c32e7f86', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[eu-west-1             64-bit  instance        karmic     ] => { :image_id => 'ami-05c2e971', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[eu-west-1             32-bit  instance        karmic     ] => { :image_id => 'ami-2fc2e95b', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
 
-      # With a value, sets the spot price to the given fraction of the
-      #   instance's full price (as found in Ironfan::Cloud::Aws::FLAVOR_INFO)
-      # With no value, returns the spot price as a fraction of the full instance price.
-      def spot_price_fraction(val=nil)
-        if val
-          spot_price( price.to_f * val )
-        else
-          spot_price / price rescue 0
-        end
-      end
+        #
+        # Lucid (Ubuntu 10.04.3)
+        #
+        %w[ap-southeast-1        64-bit  ebs             lucid      ] => { :image_id => 'ami-77f28d25', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ap-southeast-1        32-bit  ebs             lucid      ] => { :image_id => 'ami-4df28d1f', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ap-southeast-1        64-bit  instance        lucid      ] => { :image_id => 'ami-57f28d05', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ap-southeast-1        32-bit  instance        lucid      ] => { :image_id => 'ami-a5f38cf7', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[eu-west-1             64-bit  ebs             lucid      ] => { :image_id => 'ami-ab4d67df', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[eu-west-1             32-bit  ebs             lucid      ] => { :image_id => 'ami-a94d67dd', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[eu-west-1             64-bit  instance        lucid      ] => { :image_id => 'ami-a54d67d1', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[eu-west-1             32-bit  instance        lucid      ] => { :image_id => 'ami-cf4d67bb', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        #
+        %w[us-east-1             64-bit  ebs             lucid      ] => { :image_id => 'ami-4b4ba522', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[us-east-1             32-bit  ebs             lucid      ] => { :image_id => 'ami-714ba518', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[us-east-1             64-bit  instance        lucid      ] => { :image_id => 'ami-fd4aa494', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[us-east-1             32-bit  instance        lucid      ] => { :image_id => 'ami-2d4aa444', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        #
+        %w[us-west-1             64-bit  ebs             lucid      ] => { :image_id => 'ami-d197c694', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[us-west-1             32-bit  ebs             lucid      ] => { :image_id => 'ami-cb97c68e', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[us-west-1             64-bit  instance        lucid      ] => { :image_id => 'ami-c997c68c', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[us-west-1             32-bit  instance        lucid      ] => { :image_id => 'ami-c597c680', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
 
-      # EC2 User data -- DNA typically used to bootstrap the machine.
-      # @param  [Hash] value -- when present, merged with the existing user data (overriding it)
-      # @return the user_data hash
-      def user_data(hsh={})
-        @settings[:user_data].merge!(hsh.to_hash) unless hsh.empty?
-        @settings[:user_data]
-      end
+        #
+        # Maverick (Ubuntu 10.10)
+        #
+        %w[ ap-southeast-1       64-bit  ebs             maverick   ] => { :image_id => 'ami-32423c60', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ ap-southeast-1       64-bit  instance        maverick   ] => { :image_id => 'ami-12423c40', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ ap-southeast-1       32-bit  ebs             maverick   ] => { :image_id => 'ami-0c423c5e', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ ap-southeast-1       32-bit  instance        maverick   ] => { :image_id => 'ami-7c423c2e', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        #
+        %w[ eu-west-1            64-bit  ebs             maverick   ] => { :image_id => 'ami-e59ca991', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ eu-west-1            64-bit  instance        maverick   ] => { :image_id => 'ami-1b9ca96f', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ eu-west-1            32-bit  ebs             maverick   ] => { :image_id => 'ami-fb9ca98f', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ eu-west-1            32-bit  instance        maverick   ] => { :image_id => 'ami-339ca947', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        #
+        %w[ us-east-1            64-bit  ebs             maverick   ] => { :image_id => 'ami-cef405a7', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ us-east-1            64-bit  instance        maverick   ] => { :image_id => 'ami-08f40561', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ us-east-1            32-bit  ebs             maverick   ] => { :image_id => 'ami-ccf405a5', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ us-east-1            32-bit  instance        maverick   ] => { :image_id => 'ami-a6f504cf', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        #
+        %w[ us-west-1            64-bit  ebs             maverick   ] => { :image_id => 'ami-af7e2eea', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ us-west-1            64-bit  instance        maverick   ] => { :image_id => 'ami-a17e2ee4', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ us-west-1            32-bit  ebs             maverick   ] => { :image_id => 'ami-ad7e2ee8', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ us-west-1            32-bit  instance        maverick   ] => { :image_id => 'ami-957e2ed0', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
 
-      def reverse_merge!(hsh)
-        super(hsh.to_mash.compact)
-        @settings[:security_groups].reverse_merge!(hsh.security_groups) if hsh.respond_to?(:security_groups)
-        @settings[:user_data      ].reverse_merge!(hsh.user_data)       if hsh.respond_to?(:user_data)
-        self
-      end
+        #
+        # Natty (Ubuntu 11.04)
+        #
+        %w[ ap-northeast-1       32-bit  ebs             natty      ] => { :image_id => 'ami-00b10501', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ ap-northeast-1       32-bit  instance        natty      ] => { :image_id => 'ami-f0b004f1', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ ap-northeast-1       64-bit  ebs             natty      ] => { :image_id => 'ami-02b10503', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ ap-northeast-1       64-bit  instance        natty      ] => { :image_id => 'ami-fab004fb', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        #
+        %w[ ap-southeast-1       32-bit  ebs             natty      ] => { :image_id => 'ami-06255f54', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ ap-southeast-1       32-bit  instance        natty      ] => { :image_id => 'ami-72255f20', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ ap-southeast-1       64-bit  ebs             natty      ] => { :image_id => 'ami-04255f56', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ ap-southeast-1       64-bit  instance        natty      ] => { :image_id => 'ami-7a255f28', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        #
+        %w[ eu-west-1            32-bit  ebs             natty      ] => { :image_id => 'ami-a4f7c5d0', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ eu-west-1            32-bit  instance        natty      ] => { :image_id => 'ami-fef7c58a', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ eu-west-1            64-bit  ebs             natty      ] => { :image_id => 'ami-a6f7c5d2', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ eu-west-1            64-bit  instance        natty      ] => { :image_id => 'ami-c0f7c5b4', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        #
+        %w[ us-east-1            32-bit  ebs             natty      ] => { :image_id => 'ami-e358958a', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ us-east-1            32-bit  instance        natty      ] => { :image_id => 'ami-c15994a8', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ us-east-1            64-bit  ebs             natty      ] => { :image_id => 'ami-fd589594', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ us-east-1            64-bit  instance        natty      ] => { :image_id => 'ami-71589518', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        #
+        %w[ us-west-1            32-bit  ebs             natty      ] => { :image_id => 'ami-43580406', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ us-west-1            32-bit  instance        natty      ] => { :image_id => 'ami-e95f03ac', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ us-west-1            64-bit  ebs             natty      ] => { :image_id => 'ami-4d580408', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ us-west-1            64-bit  instance        natty      ] => { :image_id => 'ami-a15f03e4', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
 
-      def region(val=nil)
-        set(:region, val)
-        if    @settings[:region]        then @settings[:region]
-        elsif default_availability_zone then default_availability_zone.gsub(/^(\w+-\w+-\d)[a-z]/, '\1')
-        else  nil
-        end
-      end
+        #
+        # Cluster Compute
+        #
+        # IMAGE   ami-6d2ce204    205199409180/Globus Provision 0.4.AMI (Ubuntu 11.04 HVM)            205199409180    available       public          x86_64  machine                 ebs             hvm             xen
+        #
+        %w[ us-east-1            64-bit  ebs             natty-cc   ] => { :image_id => 'ami-6d2ce204', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
 
-      def placement_group(val=nil)
-        set(:placement_group, val)
-        @settings[:placement_group] || owner.cluster_name
-      end
+        #
+        # Oneiric (Ubuntu 11.10)
+        #
+        %w[ ap-northeast-1       32-bit  ebs             oneiric    ] => { :image_id => 'ami-84902785', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ ap-northeast-1       32-bit  instance        oneiric    ] => { :image_id => 'ami-5e90275f', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ ap-northeast-1       64-bit  ebs             oneiric    ] => { :image_id => 'ami-88902789', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ ap-northeast-1       64-bit  instance        oneiric    ] => { :image_id => 'ami-7c90277d', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        #
+        %w[ ap-southeast-1       32-bit  ebs             oneiric    ] => { :image_id => 'ami-0a327758', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ ap-southeast-1       32-bit  instance        oneiric    ] => { :image_id => 'ami-00327752', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ ap-southeast-1       64-bit  ebs             oneiric    ] => { :image_id => 'ami-0832775a', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ ap-southeast-1       64-bit  instance        oneiric    ] => { :image_id => 'ami-04327756', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        #
+        %w[ eu-west-1            32-bit  ebs             oneiric    ] => { :image_id => 'ami-11f0cc65', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ eu-west-1            32-bit  instance        oneiric    ] => { :image_id => 'ami-4ff0cc3b', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ eu-west-1            64-bit  ebs             oneiric    ] => { :image_id => 'ami-1df0cc69', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ eu-west-1            64-bit  instance        oneiric    ] => { :image_id => 'ami-23f0cc57', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        #
+        %w[ us-east-1            32-bit  ebs             oneiric    ] => { :image_id => 'ami-a562a9cc', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ us-east-1            32-bit  instance        oneiric    ] => { :image_id => 'ami-3962a950', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ us-east-1            64-bit  ebs             oneiric    ] => { :image_id => 'ami-bf62a9d6', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ us-east-1            64-bit  instance        oneiric    ] => { :image_id => 'ami-c162a9a8', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        #
+        %w[ us-west-1            32-bit  ebs             oneiric    ] => { :image_id => 'ami-c9a1fe8c', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ us-west-1            32-bit  instance        oneiric    ] => { :image_id => 'ami-21a1fe64', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ us-west-1            64-bit  ebs             oneiric    ] => { :image_id => 'ami-cba1fe8e', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ us-west-1            64-bit  instance        oneiric    ] => { :image_id => 'ami-3fa1fe7a', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        #
+        %w[ us-west-2            32-bit  ebs             oneiric    ] => { :image_id => 'ami-ea9a17da', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ us-west-2            32-bit  instance        oneiric    ] => { :image_id => 'ami-f49a17c4', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ us-west-2            64-bit  ebs             oneiric    ] => { :image_id => 'ami-ec9a17dc', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+        %w[ us-west-2            64-bit  instance        oneiric    ] => { :image_id => 'ami-fe9a17ce', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
+      })
 
-      def default_availability_zone
-        availability_zones.first if availability_zones
-      end
-
-      # Bring the ephemeral storage (local scratch disks) online
-      def mount_ephemerals(attrs={})
-        owner.volume(:ephemeral0, attrs){ device '/dev/sdb'; volume_id 'ephemeral0' ; mount_point '/mnt' ; tags( :bulk => true, :local => true, :fallback => true) } if flavor_info[:ephemeral_volumes] > 0
-        owner.volume(:ephemeral1, attrs){ device '/dev/sdc'; volume_id 'ephemeral1' ; mount_point '/mnt2'; tags( :bulk => true, :local => true, :fallback => true) } if flavor_info[:ephemeral_volumes] > 1
-        owner.volume(:ephemeral2, attrs){ device '/dev/sdd'; volume_id 'ephemeral2' ; mount_point '/mnt3'; tags( :bulk => true, :local => true, :fallback => true) } if flavor_info[:ephemeral_volumes] > 2
-        owner.volume(:ephemeral3, attrs){ device '/dev/sde'; volume_id 'ephemeral3' ; mount_point '/mnt4'; tags( :bulk => true, :local => true, :fallback => true) } if flavor_info[:ephemeral_volumes] > 3
-      end
-
-      # Utility methods
-
-      require 'pry'
-
-      def image_info
-        ec2_image_info[ [region, bits, backing, image_name] ] or ( ui.warn "Make sure to define the machine's region, bits, backing and image_name. (Have #{[region, bits, backing, image_name, virtualization].inspect})" ; {} )
-      end
-
-      def ec2_image_info
-        return @image_info_mapping if @image_info_mapping
-        @image_info_mapping = Chef::Config[:ec2_image_info]
-        # @image_info_mapping.each{|key, val| key[4] ||= 'pv' }
-        @image_info_mapping
-      end
-
-      def virtualization
-        flavor_info[:virtualization] || 'pv'
-      end
+      magic :availability_zones, Array, :default => ['us-east-1d']
+      magic :backing, String, :default => 'ebs'
+      magic :ec2_image_info, Hash, :default => Chef::Config[:ec2_image_info]
+      magic :flavor, String, :default => 't1.micro'
+      magic :image_name, String, :default => 'natty'
+      magic :keypair, Whatever
+      magic :mount_ephemerals, String
+      magic :monitoring, String
+      magic :permanent, String
+      magic :public_ip, String
+      magic :region, String, :default => lambda {
+        default_availability_zone.gsub(/^(\w+-\w+-\d)[a-z]/, '\1') if default_availability_zone
+      }
+      collection :security_groups, Ironfan::CloudDsl::SecurityGroup, :resolution => lambda {|f| merge_resolve(f) }
+      magic :ssh_user, String, :default => 'ubuntu'
+      magic :ssh_identity_dir, String, :default => Chef::Config.ec2_key_dir
+      magic :ssh_identity_file, String #, :default => "#{keypair}.pem"
+      magic :subnet, String
+      magic :user_data, Hash, :default => {}
+      magic :validation_key, String, :default => IO.read(Chef::Config.validation_key) rescue ''
+      magic :vpc, String
 
       def list_images
         ui.info("Available images:")
@@ -236,24 +212,49 @@ module Ironfan
         end
       end
 
-      def vpc(val=nil)
-        set :vpc, val
+      def default_availability_zone
+        availability_zones.first if availability_zones
       end
 
-      def subnet(val=nil)
-        set :subnet, val
+      def image_info
+        ec2_image_info[ [region, bits, backing, image_name] ] or ( ui.warn "Make sure to define the machine's region, bits, backing and image_name. (Have #{[region, bits, backing, image_name, virtualization].inspect})" ; {} )
       end
 
-      def flavor(val=nil)
-        if val && (not FLAVOR_INFO.has_key?(val.to_s))
-          ui.warn("Unknown machine image flavor '#{val}'")
-          list_flavors
-        end
-        set :flavor, val
+      # EC2 User data -- DNA typically used to bootstrap the machine.
+      # @param  [Hash] value -- when present, merged with the existing user data (overriding it)
+      # @return the user_data hash
+      def user_data(hsh={})
+        result = read_attribute(:user_data)
+        write_attribute(:user_data, result.merge!(hsh.to_hash)) unless hsh.empty?
+        result
+      end
+
+      # Sets default root volume for AWS
+      def defaults
+        owner.volume(:root).reverse_merge!({
+            :device      => '/dev/sda1',
+            :mount_point => '/',
+            :mountable   => false,
+          })
+        super
+      end
+
+      # The instance bitness, drawn from the compute flavor's info
+      def bits
+        flavor_info[:bits]
+      end
+
+      def virtualization
+        flavor_info[:virtualization] || 'pv'
       end
 
       def flavor_info
-        FLAVOR_INFO[flavor] or ( ui.warn "Please define the machine's flavor: have #{self.inspect}" ; {} )
+        if flavor && (not FLAVOR_INFO.has_key?(flavor))
+          ui.warn("Unknown machine image flavor '#{val}'")
+          list_flavors
+          return nil
+        end
+        FLAVOR_INFO[flavor]
       end
 
       def list_flavors
@@ -293,159 +294,187 @@ module Ironfan
         'cc1.8xlarge' => { :price => 2.40,  :bits => '64-bit', :ram =>  61952, :cores =>16, :core_size => 5.50, :inst_disks => 8, :inst_disk_size => 3370, :ephemeral_volumes => 4, :placement_groupable => true, :virtualization => 'hvm' },
         'cg1.4xlarge' => { :price => 2.10,  :bits => '64-bit', :ram =>  22528, :cores => 8, :core_size => 4.19, :inst_disks => 4, :inst_disk_size => 1690, :ephemeral_volumes => 2, :placement_groupable => true, :virtualization => 'hvm' },
       }
-
-      #
-      # To add to this list, use this snippet:
-      #
-      #     Chef::Config[:ec2_image_info] ||= {}
-      #     Chef::Config[:ec2_image_info].merge!({
-      #       # ... lines like the below
-      #     })
-      #
-      # in your knife.rb or whereever. We'll notice that it exists and add to it, rather than clobbering it.
-      #
-      Chef::Config[:ec2_image_info] ||= {}
-      Chef::Config[:ec2_image_info].merge!({
-
-          #
-          # Lucid (Ubuntu 9.10)
-          #
-          %w[us-east-1             64-bit  instance        karmic     ] => { :image_id => 'ami-55739e3c', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[us-east-1             32-bit  instance        karmic     ] => { :image_id => 'ami-bb709dd2', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[us-west-1             64-bit  instance        karmic     ] => { :image_id => 'ami-cb2e7f8e', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[us-west-1             32-bit  instance        karmic     ] => { :image_id => 'ami-c32e7f86', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[eu-west-1             64-bit  instance        karmic     ] => { :image_id => 'ami-05c2e971', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[eu-west-1             32-bit  instance        karmic     ] => { :image_id => 'ami-2fc2e95b', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-
-          #
-          # Lucid (Ubuntu 10.04.3)
-          #
-          %w[ap-southeast-1        64-bit  ebs             lucid      ] => { :image_id => 'ami-77f28d25', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ap-southeast-1        32-bit  ebs             lucid      ] => { :image_id => 'ami-4df28d1f', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ap-southeast-1        64-bit  instance        lucid      ] => { :image_id => 'ami-57f28d05', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ap-southeast-1        32-bit  instance        lucid      ] => { :image_id => 'ami-a5f38cf7', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[eu-west-1             64-bit  ebs             lucid      ] => { :image_id => 'ami-ab4d67df', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[eu-west-1             32-bit  ebs             lucid      ] => { :image_id => 'ami-a94d67dd', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[eu-west-1             64-bit  instance        lucid      ] => { :image_id => 'ami-a54d67d1', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[eu-west-1             32-bit  instance        lucid      ] => { :image_id => 'ami-cf4d67bb', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          #
-          %w[us-east-1             64-bit  ebs             lucid      ] => { :image_id => 'ami-4b4ba522', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[us-east-1             32-bit  ebs             lucid      ] => { :image_id => 'ami-714ba518', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[us-east-1             64-bit  instance        lucid      ] => { :image_id => 'ami-fd4aa494', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[us-east-1             32-bit  instance        lucid      ] => { :image_id => 'ami-2d4aa444', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          #
-          %w[us-west-1             64-bit  ebs             lucid      ] => { :image_id => 'ami-d197c694', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[us-west-1             32-bit  ebs             lucid      ] => { :image_id => 'ami-cb97c68e', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[us-west-1             64-bit  instance        lucid      ] => { :image_id => 'ami-c997c68c', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[us-west-1             32-bit  instance        lucid      ] => { :image_id => 'ami-c597c680', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-
-          #
-          # Maverick (Ubuntu 10.10)
-          #
-          %w[ ap-southeast-1       64-bit  ebs             maverick   ] => { :image_id => 'ami-32423c60', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ ap-southeast-1       64-bit  instance        maverick   ] => { :image_id => 'ami-12423c40', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ ap-southeast-1       32-bit  ebs             maverick   ] => { :image_id => 'ami-0c423c5e', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ ap-southeast-1       32-bit  instance        maverick   ] => { :image_id => 'ami-7c423c2e', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          #
-          %w[ eu-west-1            64-bit  ebs             maverick   ] => { :image_id => 'ami-e59ca991', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ eu-west-1            64-bit  instance        maverick   ] => { :image_id => 'ami-1b9ca96f', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ eu-west-1            32-bit  ebs             maverick   ] => { :image_id => 'ami-fb9ca98f', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ eu-west-1            32-bit  instance        maverick   ] => { :image_id => 'ami-339ca947', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          #
-          %w[ us-east-1            64-bit  ebs             maverick   ] => { :image_id => 'ami-cef405a7', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ us-east-1            64-bit  instance        maverick   ] => { :image_id => 'ami-08f40561', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ us-east-1            32-bit  ebs             maverick   ] => { :image_id => 'ami-ccf405a5', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ us-east-1            32-bit  instance        maverick   ] => { :image_id => 'ami-a6f504cf', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          #
-          %w[ us-west-1            64-bit  ebs             maverick   ] => { :image_id => 'ami-af7e2eea', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ us-west-1            64-bit  instance        maverick   ] => { :image_id => 'ami-a17e2ee4', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ us-west-1            32-bit  ebs             maverick   ] => { :image_id => 'ami-ad7e2ee8', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ us-west-1            32-bit  instance        maverick   ] => { :image_id => 'ami-957e2ed0', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-
-          #
-          # Natty (Ubuntu 11.04)
-          #
-          %w[ ap-northeast-1       32-bit  ebs             natty      ] => { :image_id => 'ami-00b10501', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ ap-northeast-1       32-bit  instance        natty      ] => { :image_id => 'ami-f0b004f1', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ ap-northeast-1       64-bit  ebs             natty      ] => { :image_id => 'ami-02b10503', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ ap-northeast-1       64-bit  instance        natty      ] => { :image_id => 'ami-fab004fb', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          #
-          %w[ ap-southeast-1       32-bit  ebs             natty      ] => { :image_id => 'ami-06255f54', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ ap-southeast-1       32-bit  instance        natty      ] => { :image_id => 'ami-72255f20', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ ap-southeast-1       64-bit  ebs             natty      ] => { :image_id => 'ami-04255f56', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ ap-southeast-1       64-bit  instance        natty      ] => { :image_id => 'ami-7a255f28', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          #
-          %w[ eu-west-1            32-bit  ebs             natty      ] => { :image_id => 'ami-a4f7c5d0', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ eu-west-1            32-bit  instance        natty      ] => { :image_id => 'ami-fef7c58a', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ eu-west-1            64-bit  ebs             natty      ] => { :image_id => 'ami-a6f7c5d2', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ eu-west-1            64-bit  instance        natty      ] => { :image_id => 'ami-c0f7c5b4', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          #
-          %w[ us-east-1            32-bit  ebs             natty      ] => { :image_id => 'ami-e358958a', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ us-east-1            32-bit  instance        natty      ] => { :image_id => 'ami-c15994a8', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ us-east-1            64-bit  ebs             natty      ] => { :image_id => 'ami-fd589594', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ us-east-1            64-bit  instance        natty      ] => { :image_id => 'ami-71589518', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          #
-          %w[ us-west-1            32-bit  ebs             natty      ] => { :image_id => 'ami-43580406', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ us-west-1            32-bit  instance        natty      ] => { :image_id => 'ami-e95f03ac', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ us-west-1            64-bit  ebs             natty      ] => { :image_id => 'ami-4d580408', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ us-west-1            64-bit  instance        natty      ] => { :image_id => 'ami-a15f03e4', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-
-          #
-          # Cluster Compute
-          #
-          # IMAGE   ami-6d2ce204    205199409180/Globus Provision 0.4.AMI (Ubuntu 11.04 HVM)            205199409180    available       public          x86_64  machine                 ebs             hvm             xen
-          #
-          %w[ us-east-1            64-bit  ebs             natty-cc   ] => { :image_id => 'ami-6d2ce204', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-
-          #
-          # Oneiric (Ubuntu 11.10)
-          #
-          %w[ ap-northeast-1       32-bit  ebs             oneiric    ] => { :image_id => 'ami-84902785', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ ap-northeast-1       32-bit  instance        oneiric    ] => { :image_id => 'ami-5e90275f', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ ap-northeast-1       64-bit  ebs             oneiric    ] => { :image_id => 'ami-88902789', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ ap-northeast-1       64-bit  instance        oneiric    ] => { :image_id => 'ami-7c90277d', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          #
-          %w[ ap-southeast-1       32-bit  ebs             oneiric    ] => { :image_id => 'ami-0a327758', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ ap-southeast-1       32-bit  instance        oneiric    ] => { :image_id => 'ami-00327752', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ ap-southeast-1       64-bit  ebs             oneiric    ] => { :image_id => 'ami-0832775a', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ ap-southeast-1       64-bit  instance        oneiric    ] => { :image_id => 'ami-04327756', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          #
-          %w[ eu-west-1            32-bit  ebs             oneiric    ] => { :image_id => 'ami-11f0cc65', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ eu-west-1            32-bit  instance        oneiric    ] => { :image_id => 'ami-4ff0cc3b', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ eu-west-1            64-bit  ebs             oneiric    ] => { :image_id => 'ami-1df0cc69', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ eu-west-1            64-bit  instance        oneiric    ] => { :image_id => 'ami-23f0cc57', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          #
-          %w[ us-east-1            32-bit  ebs             oneiric    ] => { :image_id => 'ami-a562a9cc', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ us-east-1            32-bit  instance        oneiric    ] => { :image_id => 'ami-3962a950', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ us-east-1            64-bit  ebs             oneiric    ] => { :image_id => 'ami-bf62a9d6', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ us-east-1            64-bit  instance        oneiric    ] => { :image_id => 'ami-c162a9a8', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          #
-          %w[ us-west-1            32-bit  ebs             oneiric    ] => { :image_id => 'ami-c9a1fe8c', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ us-west-1            32-bit  instance        oneiric    ] => { :image_id => 'ami-21a1fe64', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ us-west-1            64-bit  ebs             oneiric    ] => { :image_id => 'ami-cba1fe8e', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ us-west-1            64-bit  instance        oneiric    ] => { :image_id => 'ami-3fa1fe7a', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          #
-          %w[ us-west-2            32-bit  ebs             oneiric    ] => { :image_id => 'ami-ea9a17da', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ us-west-2            32-bit  instance        oneiric    ] => { :image_id => 'ami-f49a17c4', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ us-west-2            64-bit  ebs             oneiric    ] => { :image_id => 'ami-ec9a17dc', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-          %w[ us-west-2            64-bit  instance        oneiric    ] => { :image_id => 'ami-fe9a17ce', :ssh_user => 'ubuntu', :bootstrap_distro => "ubuntu10.04-gems", },
-        })
     end
 
-    class Slicehost < Base
-      # server_name
-      # slicehost_password
-      # Proc.new { |password| Chef::Config[:knife][:slicehost_password] = password }
+    class VirtualBox < Base
+      # # These fields are probably nonsense in VirtualBox
+      # magic :availability_zones, String
+      # magic :backing, String
+      # magic :default_availability_zone, String
+      # magic :flavor_info, Array, :default => { :placement_groupable => false }
+      # magic :keypair, Ironfan::PrivateKey
+      # magic :mount_ephemerals, String
+      # magic :monitoring, String
+      # magic :permanent, String
+      # magic :public_ip, String
+      # magic :security_groups, Array, :default => { :keys => nil }
+      # magic :subnet, String
+      # magic :user_data, String
+      # magic :validation_key, String
+      # magic :vpc, String
 
-      # personality
+      def initialize(*args)
+        Chef::Log.warn("Several fields (e.g. - availability_zones, backing, mount_ephemerals, etc.) are nonsense in VirtualBox context")
+        super(*args)
+      end
     end
 
-    class Rackspace < Base
-      # api_key, api_username, server_name
-    end
-
-    class Terremark < Base
-      # password, username, service
-    end
   end
 end
+
+
+# module Ironfan
+#   module Cloud
+# 
+#     #
+#     # Right now only one cloud provider is implemented, so the separation
+#     # between `cloud` and `cloud(:ec2)` is muddy.
+#     #
+#     # The goal though is to allow
+#     #
+#     # * cloud with no predicate -- definitions that apply to all cloud
+#     #   providers. If you only use one provider ever nothing stops you from
+#     #   always saying `cloud`.
+#     # * Declarations irrelevant to other providers are acceptable and will be ignored
+#     # * Declarations that are wrong in the context of other providers (a `public_ip`
+#     #   that is not available) will presumably cause a downstream error -- it's
+#     #   your responsibility to overlay with provider-correct values.
+#     # * There are several declarations that *could* be sensibly abstracted, but
+#     #   are not. Rather than specifying `flavor 'm1.xlarge'`, I could ask for
+#     #   :ram => 15, :cores => 4 or storage => 1500 and get the cheapest machine
+#     #   that met or exceeded each constraint -- the default of `:price =>
+#     #   :smallest` would get me a t1.micro on EC2, a 256MB on
+#     #   Rackspace. Availability zones could also plausibly be parameterized.
+#     #
+#     # @example
+#     #     # these apply regardless of cloud provider
+#     #     cloud do
+#     #       # this makes sense everywhere
+#     #       image_name            'maverick'
+#     #
+#     #       # this is not offered by many providers, and its value is non-portable;
+#     #       # but if you only run in one cloud there's harm in putting it here
+#     #       # or overriding it.
+#     #       public_ip             '1.2.3.4'
+#     #
+#     #       # Implemented differently across providers but its meaning is clear
+#     #       security_group        :nagios
+#     #
+#     #       # This is harmless for the other clouds
+#     #       availability_zones   ['us-east-1d']
+#     #     end
+#     #
+#     #     # these only apply to ec2 launches.
+#     #     # `ec2` is sugar for `cloud(:ec2)`.
+#     #     ec2 do
+#     #       spot_price_fraction   0.4
+#     #     end
+#     #
+#     class Base < Ironfan::DslObject
+#       has_keys(
+#         :name, :flavor, :image_name, :image_id, :keypair,
+#         :chef_client_script, :public_ip, :permanent )
+#       attr_accessor :owner
+# 
+#       def initialize(owner, *args)
+#         self.owner = owner
+#         super(*args)
+#       end
+# 
+#       def validation_key
+#         IO.read(Chef::Config.validation_key) rescue ''
+#       end
+# 
+#       # The instance price, drawn from the compute flavor's info
+#       def price
+#         flavor_info[:price]
+#       end
+# 
+#     protected
+#       # If value was explicitly set, use that; if the Chef::Config[:ec2_image_info] implies a value use that; otherwise use the default
+#       def from_setting_or_image_info(key, val=nil, default=nil)
+#         @settings[key] = val unless val.nil?
+#         return @settings[key]  if @settings.include?(key)
+#         return image_info[key] unless image_info.nil?
+#         return default       # otherwise
+#       end
+#     end
+# 
+#     class Ec2 < Base
+#       has_keys(
+#         :region, :availability_zones, :backing,
+#         :spot_price, :spot_price_fraction,
+#         :user_data, :security_groups,
+#         :monitoring, :placement_group,
+#         :vpc, :subnet
+#         )
+# 
+#       def initialize(*args)
+#         super *args
+#         name :ec2 # cloud provider name
+#         @settings[:security_groups]      ||= Mash.new
+#         @settings[:user_data]            ||= Mash.new
+#       end
+# 
+# 
+#       # With a value, sets the spot price to the given fraction of the
+#       #   instance's full price (as found in Ironfan::Cloud::Aws::FLAVOR_INFO)
+#       # With no value, returns the spot price as a fraction of the full instance price.
+#       def spot_price_fraction(val=nil)
+#         if val
+#           spot_price( price.to_f * val )
+#         else
+#           spot_price / price rescue 0
+#         end
+#       end
+# 
+#       def reverse_merge!(hsh)
+#         super(hsh.to_mash.compact)
+#         @settings[:security_groups].reverse_merge!(hsh.security_groups) if hsh.respond_to?(:security_groups)
+#         @settings[:user_data      ].reverse_merge!(hsh.user_data)       if hsh.respond_to?(:user_data)
+#         self
+#       end
+# 
+#       def region(val=nil)
+#         set(:region, val)
+#         if    @settings[:region]        then @settings[:region]
+#         elsif default_availability_zone then default_availability_zone.gsub(/^(\w+-\w+-\d)[a-z]/, '\1')
+#         else  nil
+#         end
+#       end
+# 
+#       def placement_group(val=nil)
+#         set(:placement_group, val)
+#         @settings[:placement_group] || owner.cluster_name
+#       end
+# 
+#       # Bring the ephemeral storage (local scratch disks) online
+#       def mount_ephemerals(attrs={})
+#         owner.volume(:ephemeral0, attrs){ device '/dev/sdb'; volume_id 'ephemeral0' ; mount_point '/mnt' ; tags( :bulk => true, :local => true, :fallback => true) } if flavor_info[:ephemeral_volumes] > 0
+#         owner.volume(:ephemeral1, attrs){ device '/dev/sdc'; volume_id 'ephemeral1' ; mount_point '/mnt2'; tags( :bulk => true, :local => true, :fallback => true) } if flavor_info[:ephemeral_volumes] > 1
+#         owner.volume(:ephemeral2, attrs){ device '/dev/sdd'; volume_id 'ephemeral2' ; mount_point '/mnt3'; tags( :bulk => true, :local => true, :fallback => true) } if flavor_info[:ephemeral_volumes] > 2
+#         owner.volume(:ephemeral3, attrs){ device '/dev/sde'; volume_id 'ephemeral3' ; mount_point '/mnt4'; tags( :bulk => true, :local => true, :fallback => true) } if flavor_info[:ephemeral_volumes] > 3
+#       end
+# 
+#       # Utility methods
+# 
+#       require 'pry'
+#     end
+# 
+#     class Slicehost < Base
+#       # server_name
+#       # slicehost_password
+#       # Proc.new { |password| Chef::Config[:knife][:slicehost_password] = password }
+# 
+#       # personality
+#     end
+# 
+#     class Rackspace < Base
+#       # api_key, api_username, server_name
+#     end
+# 
+#     class Terremark < Base
+#       # password, username, service
+#     end
+#   end
+# end
